Move Black into a plugin to enable turning it on and off
