+++ pants.ini
@@ -46,10 +46,6 @@ pythonpath: [
 
 backend_packages: +[
     "pants.backend.docgen",
-    "internal_backend.rules_for_testing",
-    "internal_backend.repositories",
-    "internal_backend.sitegen",
-    "internal_backend.utilities",
     "pants.contrib.avro",
     "pants.contrib.awslambda.python",
     "pants.contrib.buildrefactor",
@@ -66,6 +62,10 @@ backend_packages: +[
     "pants.contrib.python.checks",
     "pants.contrib.scrooge",
     "pants.contrib.thrifty",
+    "internal_backend.rules_for_testing",
+    "internal_backend.repositories",
+    "internal_backend.sitegen",
+    "internal_backend.utilities",
   ]
 
 # The pants script in this repo consumes these files to run pants
@@ -82,8 +82,6 @@ pants_ignore: +[
     '/build-support/bin/native/src',
   ]
 
-[black]
-config: pyproject.toml
 
 [cache]
 # Caching is on globally by default, but we disable it here for development purposes.
