+++ src/python/pants/backend/python/lint/black/rules.py
@@ -6,13 +6,13 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional, Tuple
 
+from pants.backend.python.lint.black.subsystem import Black
 from pants.backend.python.rules.pex import (
   CreatePex,
   Pex,
   PexInterpreterConstraints,
   PexRequirements,
 )
-from pants.backend.python.subsystems.black import Black
 from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.subsystems.subprocess_environment import SubprocessEncodingEnvironment
 from pants.backend.python.targets.formattable_python_target import FormattablePythonTarget
