+++ src/python/pants/backend/python/register.py
@@ -6,7 +6,6 @@ from pants.backend.python.python_artifact import PythonArtifact
 from pants.backend.python.python_requirement import PythonRequirement
 from pants.backend.python.python_requirements import PythonRequirements
 from pants.backend.python.rules import (
-  black,
   download_pex_bin,
   inject_init,
   pex,
@@ -14,7 +13,6 @@ from pants.backend.python.rules import (
   python_test_runner,
 )
 from pants.backend.python.subsystems import python_native_code, subprocess_environment
-from pants.backend.python.targets import formattable_python_target
 from pants.backend.python.targets.python_app import PythonApp
 from pants.backend.python.targets.python_binary import PythonBinary
 from pants.backend.python.targets.python_distribution import PythonDistribution
@@ -94,9 +92,7 @@ def register_goals():
 
 def rules():
   return (
-    *black.rules(),
     *download_pex_bin.rules(),
-    *formattable_python_target.rules(),
     *inject_init.rules(),
     *pex.rules(),
     *python_test_runner.rules(),
