+++ src/python/pants/backend/python/rules/BUILD
@@ -43,20 +43,9 @@ python_tests(
   name='integration',
   sources=globs('*_integration_test.py'),
   dependencies=[
-    ':rules',
-    'src/python/pants/backend/python/subsystems',
-    'src/python/pants/backend/python/targets',
-    'src/python/pants/engine:fs',
-    'src/python/pants/engine:rules',
-    'src/python/pants/engine:selectors',
-    'src/python/pants/engine/legacy:structs',
-    'src/python/pants/rules/core',
-    'src/python/pants/source',
-    'src/python/pants/testutil:test_base',
-    'src/python/pants/testutil/subsystem',
     'src/python/pants/testutil:int-test',
     'testprojects/tests/python/pants:dummies_directory',
   ],
   tags = {'integration'},
-  timeout = 160,
+  timeout = 90,
 )
